module_09 ex01 using stack of double instead of int

+++ cpp_module_09/ex01/RPN.cpp
@@ -33,9 +33,9 @@ void	RPN::setExpression(const String& expression) {
 bool	RPN::runOperation(char operation) {
 	if (stack.size() <= 1)
 		return false;
-	int o2 = stack.top();
+	double o2 = stack.top();
 	stack.pop();
-	int o1 = stack.top();
+	double o1 = stack.top();
 	stack.pop();
 
 	switch(operation) {
